Minor cleanup

+++ client/src/main/java/io/confluent/kafka/schemaregistry/AbstractSchemaProvider.java
@@ -105,7 +105,7 @@ public abstract class AbstractSchemaProvider implements SchemaProvider {
       // This handles the case where current schema is without confluent:version
       return newSchema.deepEquals(newPrev);
     } else if (schemaVer != null && prevVer == null) {
-      if (!Objects.equals(schemaVer.intValue(), prev.version())) {
+      if (!Objects.equals(schemaVer, prev.version())) {
         // The incoming confluent:version must match the actual version of the prev schema
         return false;
       }
